requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
imbalanced-learn==0.11.0
joblib==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Deep Learning (Optional)
tensorflow==2.15.0

# API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Data Validation
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0

# Notebook
jupyter==1.0.0
ipykernel==6.28.0
ipywidgets==8.1.1

# Monitoring & Logging
loguru==0.7.2

# Data Download
kaggle==1.5.16
openpyxl==3.1.2